% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiscape-package.R
\docType{package}
\name{emojiscape-package}
\alias{emojiscape}
\alias{emojiscape-package}
\title{emojiscape: Randomised Emoji Scenes}
\description{
Print to the console a randomised scene composed of emoji.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
