% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiscape.R
\name{get_set}
\alias{get_set}
\title{See Emoji In Each Terrain}
\usage{
get_set(terrain)
}
\arguments{
\item{terrain}{Character. What type of locale? See 'details' below for
options.}
}
\value{
A data.frame of 4 columns and 3 rows.
}
\description{
Print a dataframe with the emoji that are found in the name terrain and their
probability slot.
}
\details{
Currently, terrain options are  "arable", "city", "desert", "forest",
"garden", "liminal", "mountains", "ocean", "pastoral", "polar", "rainforest",
"sky", "space", "suburbs", "traffic", "undergrowth" and "woods".
}
\examples{
get_set("space")
}
