% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiscape.R
\name{generate}
\alias{generate}
\title{Generate An Emoji Scene}
\usage{
generate(terrain = c("ocean", "city", "desert"), size = 10)
}
\arguments{
\item{terrain}{Character. What type of locale?}

\item{size}{Numeric. Width of the emoji matrix.}
}
\value{
Class NULL. Matrix of emoji printed to the console line by line with
cat().
}
\description{
Print to the console a randomised scene composed of emoji.
}
\examples{
generate(terrain = "ocean")
}
