% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojiscape.R
\name{generate}
\alias{generate}
\title{Generate An Emoji Scene}
\usage{
generate(
  terrain = c("arable", "city", "desert", "forest", "garden", "liminal", "mountains",
    "ocean", "pastoral", "polar", "rainforest", "sky", "space", "suburbs", "traffic",
    "undergrowth", "woods"),
  grid_size = 10,
  prob_common = 0.75,
  prob_uncommon = 0.2,
  prob_rare = 0.05
)
}
\arguments{
\item{terrain}{Character. What type of locale? See 'details' below for
options.}

\item{grid_size}{Numeric, greater than zero. Width and length of the emoji
matrix.}

\item{prob_common}{Numeric. Probability of occurrence for the
'common' emoji in the terrain set.}

\item{prob_uncommon}{Numeric. Probability of occurrence for the
'uncommon' emoji in the terrain set.}

\item{prob_rare}{Numeric. Probability of occurrence for the
'rare' emoji in the terrain set.}
}
\value{
Class NULL. Matrix of emoji printed to the console line by line with
cat().
}
\description{
Print to the console a randomised scene composed of emoji, given a specific
terrain type.
}
\details{
Your terrain choice results in the selection of a predetermined three-
emoji set, from which a sample is taken to fill a square matrix with
dimensions of grid_size.

Currently, terrain options are "arable", "city", "desert", "forest",
"garden", "liminal", "mountains", "ocean", "pastoral", "polar", "rainforest",
"roads", "sky", "space", "undergrowth" and "woods".

The prob_* arguments are passed to sample() and are relative
to each other.
}
\examples{
generate("ocean")
}
